Base channel message type
