+++ src/daf/message/text_based.py
@@ -35,7 +35,7 @@ sql.register_type("MessageMODE", "clear-send")
 @instance_track.track_id
 @doc.doc_category("Messages", path="message")
 @sql.register_type("MessageTYPE")
-class TextMESSAGE(BaseMESSAGE):
+class TextMESSAGE(BaseChannelMessage):
     """
     This class is used for creating objects that represent messages which will be sent to Discord's TEXT CHANNELS.
 
@@ -115,23 +115,23 @@ class TextMESSAGE(BaseMESSAGE):
     """
 
     __slots__ = (
-        "channels",
         "mode",
         "sent_messages",
     )
 
     @typechecked
-    def __init__(self,
-                 start_period: Union[timedelta, int, None],
-                 end_period: Union[int, timedelta],
-                 data: Union[Iterable[Union[str, discord.Embed, FILE]], str, discord.Embed, FILE, _FunctionBaseCLASS],
-                 channels: Union[Iterable[Union[int, discord.TextChannel, discord.Thread]], AutoCHANNEL],
-                 mode: Optional[Literal["send", "edit", "clear-send"]] = "send",
-                 start_in: Optional[Union[timedelta, datetime]] = timedelta(seconds=0),
-                 remove_after: Optional[Union[int, timedelta, datetime]] = None):
-        super().__init__(start_period, end_period, data, start_in, remove_after)
+    def __init__(
+        self,
+        start_period: Union[timedelta, int, None],
+        end_period: Union[int, timedelta],
+        data: Union[Iterable[Union[str, discord.Embed, FILE]], str, discord.Embed, FILE, _FunctionBaseCLASS],
+        channels: Union[Iterable[Union[int, discord.TextChannel, discord.Thread]], AutoCHANNEL],
+        mode: Optional[Literal["send", "edit", "clear-send"]] = "send",
+        start_in: Optional[Union[timedelta, datetime]] = timedelta(seconds=0),
+        remove_after: Optional[Union[int, timedelta, datetime]] = None
+    ):
+        super().__init__(start_period, end_period, data, channels, start_in, remove_after)
         self.mode = mode
-        self.channels = channels
         # Dictionary for storing last sent message for each channel
         self.sent_messages: Dict[int, discord.Message] = {}
 
@@ -150,37 +150,6 @@ class TextMESSAGE(BaseMESSAGE):
             seconds += 10
 
         return timedelta(seconds=seconds)
-
-    def _check_state(self) -> bool:
-        """
-        Checks if the message is ready to be deleted.
-
-        Returns
-        ----------
-        True
-            The message should be deleted.
-        False
-            The message is in proper state, do not delete.
-        """
-        return super()._check_state() or not bool(self.channels)
-
-    def _update_state(self, err_channels: List[dict]):
-        "Updates the state of the object based on errored channels."
-        super()._update_state()
-        # Remove any channels that returned with code status 404 (They no longer exist)
-        for data in err_channels:
-            reason = data["reason"]
-            channel = data["channel"]
-            if isinstance(reason, discord.HTTPException):
-                if (
-                    reason.status == 403 or  # Forbidden
-                    reason.code in {50007, 10003}  # Not Forbidden, but bad error codes
-                ):
-                    trace(
-                        f"Removing channel '{channel.name}' ({channel.id}) [Guild '{channel.guild.name}' ({channel.guild.id})], because user '{self.parent.parent.client.user.name}' lacks permissions.",
-                        TraceLEVELS.WARNING
-                    )
-                    self.channels.remove(channel)
 
     def _check_period(self):
         """
@@ -329,44 +298,7 @@ class TextMESSAGE(BaseMESSAGE):
         ValueError
             No valid channels were passed to object"
         """
-        if parent is None:
-            return
-
-        ch_i = 0
-        self.parent = parent
-        cl = parent.parent.client
-        _guild = parent.apiobject
-        to_remove = []
-
-        if isinstance(self.channels, AutoCHANNEL):
-            await self.channels.initialize(self, "text_channels")
-        else:
-            self.channels: List[Union[discord.TextChannel, discord.Thread]]
-            for ch_i, channel in enumerate(self.channels):
-                if isinstance(channel, discord.abc.GuildChannel):
-                    channel_id = channel.id
-                else:
-                    # Snowflake IDs provided
-                    channel_id = channel
-                    channel = self.channels[ch_i] = cl.get_channel(channel_id)
-
-                if channel is None:
-                    trace(f"Unable to get channel from ID {channel_id}", TraceLEVELS.WARNING)
-                    to_remove.append(channel)
-                elif type(channel) not in {discord.TextChannel, discord.Thread}:
-                    raise TypeError(f"TextMESSAGE object received invalid channel type of {type(channel).__name__}")
-                elif channel.guild != _guild:
-                    raise ValueError(
-                        f"{channel.name}(ID: {channel_id}) not in {_guild.name}(ID: {_guild.id}), "
-                        f"but is part of {channel.guild.name}(ID: {channel.guild.id})"
-                    )
-
-            for channel in to_remove:
-                self.channels.remove(channel)
-
-            if not self.channels:
-                raise ValueError(f"No valid channels were passed to {self} object")
-
+        await super().initialize(parent, {discord.TextChannel, discord.Thread}, lambda: parent.apiobject.text_channels)
         # Increase period to slow mode delay if it is lower
         self._check_period()
 
@@ -513,93 +445,6 @@ class TextMESSAGE(BaseMESSAGE):
                 handled, action = await self._handle_error(channel, ex)
                 if not handled:
                     return {"success": False, "reason": ex, "action": action}
-
-    @async_util.with_semaphore("update_semaphore")
-    async def _send(self) -> MessageSendResult:
-        """
-        Sends the data into the channels.
-
-        Returns
-        ----------
-        MessageSendResult
-            The result of message send attempt.
-        """
-        # Acquire mutex to prevent update method from writing while sending
-        data_to_send = await self._get_data()
-        if any(data_to_send.values()):  # There is data to be send
-            errored_channels = []
-            succeeded_channels = []
-
-            # Send to channels
-            for channel in self.channels:
-                # Clear previous messages sent to channel if mode is MODE_DELETE_SEND
-                context = await self._send_channel(channel, **data_to_send)
-                if context["success"]:
-                    succeeded_channels.append(channel)
-                else:
-                    errored_channels.append({"channel": channel, "reason": context["reason"]})
-                    action = context["action"]
-                    if action is ChannelErrorAction.SKIP_CHANNELS:  # Don't try to send to other channels
-                        break
-
-                    elif action is ChannelErrorAction.REMOVE_ACCOUNT:
-                        return MessageSendResult(
-                            self.generate_log_context(
-                                **data_to_send, succeeded_ch=succeeded_channels, failed_ch=errored_channels
-                            ),
-                            MSG_SEND_STATUS_ERROR_REMOVE_ACCOUNT
-                        )
-
-            self._update_state(errored_channels)
-            return MessageSendResult(
-                self.generate_log_context(**data_to_send, succeeded_ch=succeeded_channels, failed_ch=errored_channels),
-                MSG_SEND_STATUS_SUCCESS if succeeded_channels or errored_channels else MSG_SEND_STATUS_NO_MESSAGE_SENT
-            )
-
-        return MessageSendResult(None, MSG_SEND_STATUS_NO_MESSAGE_SENT)
-
-    @typechecked
-    @async_util.with_semaphore("update_semaphore")
-    async def update(self, _init_options: Optional[dict] = None, **kwargs: Any):
-        """
-        .. versionadded:: v2.0
-
-        Used for changing the initialization parameters the object was initialized with.
-
-        .. warning::
-            Upon updating, the internal state of objects get's reset,
-            meaning you basically have a brand new created object.
-
-        Parameters
-        -------------
-        **kwargs: Any
-            Custom number of keyword parameters which you want to update,
-            these can be anything that is available during the object creation.
-
-        Raises
-        -----------
-        TypeError
-            Invalid keyword argument was passed
-        Other
-            Raised from .initialize() method.
-        """
-        if "start_in" not in kwargs:
-            # This parameter does not appear as attribute, manual setting necessary
-            kwargs["start_in"] = self.next_send_time
-
-        if "data" not in kwargs:
-            kwargs["data"] = self._data
-
-        kwargs["channels"] = channels = kwargs.get("channels", self.channels)
-        if isinstance(channels, AutoCHANNEL):
-            await channels.update(init_options={"parent": self, "channel_type": "text_channels"})
-        elif not isinstance(self.channels[0], int):  # Not initialized (newly created):
-            kwargs["channels"] = [x.id for x in self.channels]
-
-        if _init_options is None:
-            _init_options = {"parent": self.parent}
-
-        await async_util.update_obj_param(self, init_options=_init_options, **kwargs)
 
 
 @instance_track.track_id
